@@ -43,55 +47,56 @@
         "is-valid-fodselsnummer": "^3.0.5",
         "lodash": "^4.17.21",
         "morgan": "^1.10.0",
-        "nav-datovelger": "^10.1.2",
-        "nav-frontend-alertstriper": "^3.0.38",
-        "nav-frontend-alertstriper-style": "^2.0.30",
-        "nav-frontend-chevron": "^1.0.24",
-        "nav-frontend-chevron-style": "^1.0.2",
-        "nav-frontend-core": "^5.0.10",
-        "nav-frontend-ekspanderbartpanel": "^3.0.45",
-        "nav-frontend-ekspanderbartpanel-style": "^1.0.33",
-        "nav-frontend-etiketter": "^2.0.21",
-        "nav-frontend-etiketter-style": "^1.0.14",
-        "nav-frontend-grid": "^1.0.44",
-        "nav-frontend-grid-style": "^0.2.25",
-        "nav-frontend-hjelpetekst": "^2.0.47",
-        "nav-frontend-hjelpetekst-style": "^2.0.36",
-        "nav-frontend-ikoner-assets": "^2.0.8",
-        "nav-frontend-js-utils": "^1.0.15",
-        "nav-frontend-knapper": "^2.0.24",
-        "nav-frontend-knapper-style": "^1.0.11",
-        "nav-frontend-lenkepanel": "^1.0.64",
-        "nav-frontend-lenkepanel-style": "^0.3.44",
-        "nav-frontend-lenker": "^1.0.53",
-        "nav-frontend-lenker-style": "^0.2.39",
-        "nav-frontend-lukknapp": "^1.0.43",
-        "nav-frontend-lukknapp-style": "^0.2.34",
-        "nav-frontend-modal": "^1.0.55",
-        "nav-frontend-modal-style": "^0.3.46",
-        "nav-frontend-paneler-style": "^0.3.31",
-        "nav-frontend-popover": "^0.0.41",
-        "nav-frontend-popover-style": "^0.0.13",
-        "nav-frontend-skjema": "^3.0.40",
-        "nav-frontend-skjema-style": "^2.0.24",
-        "nav-frontend-spinner": "^2.0.12",
-        "nav-frontend-spinner-style": "^0.2.9",
-        "nav-frontend-stegindikator": "^0.0.65",
-        "nav-frontend-stegindikator-style": "^1.0.41",
-        "nav-frontend-typografi": "^2.0.37",
-        "nav-frontend-typografi-style": "^1.0.31",
-        "nav-frontend-veileder": "^2.0.37",
-        "nav-frontend-veileder-style": "^2.0.27",
-        "nav-frontend-veilederpanel": "^0.1.46",
-        "nav-frontend-veilederpanel-style": "0.0.40",
-        "react": "^17.0.1",
+        "nav-datovelger": "^12.6.0",
+        "nav-frontend-alertstriper": "^4.0.2",
+        "nav-frontend-alertstriper-style": "^3.0.2",
+        "nav-frontend-chevron": "^1.0.30",
+        "nav-frontend-chevron-style": "^1.0.4",
+        "nav-frontend-core": "^6.0.1",
+        "nav-frontend-ekspanderbartpanel": "^4.0.4",
+        "nav-frontend-ekspanderbartpanel-style": "^2.0.2",
+        "nav-frontend-etiketter": "^3.0.3",
+        "nav-frontend-etiketter-style": "^2.0.3",
+        "nav-frontend-grid": "^2.0.2",
+        "nav-frontend-grid-style": "^1.0.3",
+        "nav-frontend-hjelpetekst": "^3.1.3",
+        "nav-frontend-hjelpetekst-style": "^3.0.2",
+        "nav-frontend-ikoner-assets": "^3.0.1",
+        "nav-frontend-js-utils": "^1.0.20",
+        "nav-frontend-knapper": "^3.1.3",
+        "nav-frontend-knapper-style": "^2.1.2",
+        "nav-frontend-lenkepanel": "^2.0.2",
+        "nav-frontend-lenkepanel-style": "^2.0.2",
+        "nav-frontend-lenker": "^2.0.2",
+        "nav-frontend-lenker-style": "^2.0.2",
+        "nav-frontend-lukknapp": "^2.0.2",
+        "nav-frontend-lukknapp-style": "^2.0.2",
+        "nav-frontend-modal": "^2.0.2",
+        "nav-frontend-modal-style": "^2.0.2",
+        "nav-frontend-paneler": "^3.0.2",
+        "nav-frontend-paneler-style": "^2.0.2",
+        "nav-frontend-popover": "^2.0.4",
+        "nav-frontend-popover-style": "^2.0.2",
+        "nav-frontend-skjema": "^4.0.6",
+        "nav-frontend-skjema-style": "^3.0.3",
+        "nav-frontend-spinner": "^3.0.1",
+        "nav-frontend-spinner-style": "^1.0.2",
+        "nav-frontend-stegindikator": "^2.0.3",
+        "nav-frontend-stegindikator-style": "^2.0.2",
+        "nav-frontend-typografi": "^4.0.2",
+        "nav-frontend-typografi-style": "^2.0.2",
+        "nav-frontend-veileder": "^3.0.3",
+        "nav-frontend-veileder-style": "^3.0.2",
+        "nav-frontend-veilederpanel": "^2.0.3",
+        "nav-frontend-veilederpanel-style": "2.0.2",
+        "react": "^18.2.0",
         "react-aria-menubutton": "^7.0.0",
         "react-day-picker": "^7.4.8",
         "react-document-title": "^2.0.3",
-        "react-dom": "^17.0.1",
-        "react-intl": "^5.10.6",
+        "react-dom": "^18.2.0",
+        "react-intl": "^6.2.5",
         "react-modal": "^3.12.1",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.4.5",
         "react-transition-group": "^4.4.1",
         "request": "^2.88.2"
     },
@@ -140,22 +146,24 @@
         "postcss-loader": "^3.0.0",
         "promise": "^8.1.0",
         "react-scripts": "^4.0.1",
+        "sass": "^1.54.9",
+        "sass-loader": "^13.0.2",
         "sinon": "^9.2.0",
         "source-map-loader": "^1.1.1",
-        "style-loader": "^2.0.0",
-        "svg-sprite-loader": "^4.1.6",
-        "terser-webpack-plugin": "^4.2.2",
+        "style-loader": "^3.3.1",
+        "svg-sprite-loader": "^6.0.11",
+        "terser-webpack-plugin": "^5.3.6",
         "ts-jest": "^24.3.0",
-        "ts-loader": "^8.0.7",
+        "ts-loader": "^9.4.2",
         "tslint": "^5.20.0",
         "tslint-config-prettier": "^1.17.0",
         "tslint-loader": "^3.5.4",
         "tslint-react": "^4.1.0",
-        "typescript": "^4.1.3",
+        "typescript": "^4.9.4",
         "uglifyjs-webpack-plugin": "^2.2.0",
         "wait-on": "^5.2.0",
-        "webpack": "^4.44.2",
-        "webpack-cli": "^3.3.9",
-        "webpack-dev-server": "^3.8.1"
+        "webpack": "^5.75.0",
+        "webpack-cli": "^5.0.1",
+        "webpack-dev-server": "^4.11.1"
     }
 }
