@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Attachment } from 'common/storage/attachment/types/Attachment';
-import { EtikettLiten } from 'nav-frontend-typografi';
+import { Element } from 'nav-frontend-typografi';
 import { useIntl } from 'react-intl';
 import { isAttachmentWithError } from 'common/storage/attachment/components/util';
 import AttachmentList from 'common/storage/attachment/components/AttachmentList';
@@ -43,11 +43,11 @@ const OmBarnetOppsummering: React.FunctionComponent<Props> = ({ barn }) => {
                         text={formatDate(barn.adopsjonsdato!)}
                     />
 
-                    <EtikettLiten className="textWithLabel__label">
+                    <Element className="textWithLabel__label">
                         {barn.fødselsdatoer.length > 1
                             ? getMessage(intl, 'oppsummering.text.medFødselsdatoer')
                             : getMessage(intl, 'oppsummering.text.medFødselsdato')}
-                    </EtikettLiten>
+                    </Element>
                     <Block margin="l" padBottom="l">
                         {barn.fødselsdatoer.map((_, index) => {
                             return (
@@ -62,9 +62,9 @@ const OmBarnetOppsummering: React.FunctionComponent<Props> = ({ barn }) => {
             )}
             {barn.adopsjonAvEktefellesBarn !== YesOrNo.UNANSWERED && (
                 <div className="oppsummering__attachments">
-                    <EtikettLiten className="textWithLabel__label">
+                    <Element className="textWithLabel__label">
                         {getMessage(intl, 'oppsummering.text.vedlagtAdopsjonBekreftelse')}
-                    </EtikettLiten>
+                    </Element>
                     <AttachmentList
                         attachments={barn.omsorgsovertakelse.filter((a: Attachment) => !isAttachmentWithError(a))}
                     />
@@ -83,9 +83,9 @@ const OmBarnetOppsummering: React.FunctionComponent<Props> = ({ barn }) => {
                         text={formatDate(barn.termindato)}
                     />
                     <div className="oppsummering__attachments">
-                        <EtikettLiten className="textWithLabel__label">
+                        <Element className="textWithLabel__label">
                             {getMessage(intl, 'oppsummering.text.vedlagtTerminbekreftelse')}
-                        </EtikettLiten>
+                        </Element>
                         <AttachmentList
                             attachments={barn.terminbekreftelse.filter((a: Attachment) => !isAttachmentWithError(a))}
                         />
