@@ -1,6 +1,7 @@
 import React, { ReactNode } from 'react';
 import Ekspanderbartpanel from 'nav-frontend-ekspanderbartpanel';
 import { Block } from '@navikt/fp-common';
+import { Undertittel } from 'nav-frontend-typografi';
 
 interface Props {
     tittel: string;
@@ -9,7 +10,7 @@ interface Props {
 
 const Oppsummeringspunkt = ({ tittel, children }: Props) => (
     <Block margin="xl">
-        <Ekspanderbartpanel apen={false} tittel={tittel}>
+        <Ekspanderbartpanel apen={false} tittel={<Undertittel>{tittel}</Undertittel>}>
             {children}
         </Ekspanderbartpanel>
     </Block>
