@@ -11,21 +11,10 @@ interface Props {
     className?: string;
 }
 
-const CustomSVGFromSprite: React.StatelessComponent<Props> = ({
-    iconRef,
-    size,
-    className,
-    ...other
-}) => {
+const CustomSVGFromSprite: React.FunctionComponent<Props> = ({ iconRef, size, className, ...other }) => {
     const viewBox = { 'view-box': iconRef.viewBox };
     return (
-        <svg
-            className={className}
-            height={size}
-            width={size}
-            {...viewBox}
-            {...other}
-        >
+        <svg className={className} height={size} width={size} {...viewBox} {...other}>
             <use xlinkHref={`#${iconRef.id}`} />
         </svg>
     );
